+++ InnerEye/ML/Histopathology/models/deepmil.py
@@ -46,6 +46,7 @@ class DeepMILModule(LightningModule):
                  pooling_layer: Callable[[int, int, int], nn.Module],
                  pool_hidden_dim: int = 128,
                  pool_out_dim: int = 1,
+                 dropout_rate: Optional[float] = None,
                  class_weights: Optional[Tensor] = None,
                  l_rate: float = 5e-4,
                  weight_decay: float = 1e-4,
@@ -64,6 +65,7 @@ class DeepMILModule(LightningModule):
         `torch.nn.Module` constructor accepting input, hidden, and output pooling `int` dimensions.
         :param pool_hidden_dim: Hidden dimension of pooling layer (default=128).
         :param pool_out_dim: Output dimension of pooling layer (default=1).
+        :param dropout_rate: Rate of pre-classifier dropout (0-1). `None` for no dropout (default).
         :param class_weights: Tensor containing class weights (default=None).
         :param l_rate: Optimiser learning rate.
         :param weight_decay: Weight decay parameter for L2 regularisation.
@@ -82,6 +84,7 @@ class DeepMILModule(LightningModule):
         self.pool_hidden_dim = pool_hidden_dim
         self.pool_out_dim = pool_out_dim
         self.pooling_layer = pooling_layer
+        self.dropout_rate = dropout_rate
         self.class_weights = class_weights
         self.encoder = encoder
         self.num_encoding = self.encoder.num_encoding
@@ -130,8 +133,14 @@ class DeepMILModule(LightningModule):
         return pooling_layer, num_features
 
     def get_classifier(self) -> Callable:
-        return nn.Linear(in_features=self.num_pooling,
-                         out_features=self.n_classes)
+        classifier_layer = nn.Linear(in_features=self.num_pooling,
+                                     out_features=self.n_classes)
+        if self.dropout_rate is None:
+            return classifier_layer
+        elif 0 <= self.dropout_rate < 1:
+            return nn.Sequential(nn.Dropout(self.dropout_rate), classifier_layer)
+        else:
+            raise ValueError(f"Dropout rate should be in [0, 1), got {self.dropout_rate}")
 
     def get_loss(self) -> Callable:
         if self.n_classes > 1:
@@ -186,13 +195,13 @@ class DeepMILModule(LightningModule):
             else:
                 log_on_epoch(self, f'{stage}/{metric_name}', metric_object)
 
-    def forward(self, images: Tensor) -> Tuple[Tensor, Tensor]:  # type: ignore
+    def forward(self, instances: Tensor) -> Tuple[Tensor, Tensor]:  # type: ignore
         with no_grad():
-            H = self.encoder(images)                        # N X L x 1 x 1
-        A, M = self.aggregation_fn(H)                       # A: K x N | M: K x L
-        M = M.view(-1, self.num_encoding * self.pool_out_dim)
-        Y_prob = self.classifier_fn(M)
-        return Y_prob, A
+            instance_features = self.encoder(instances)                    # N X L x 1 x 1
+        attentions, bag_features = self.aggregation_fn(instance_features)  # K x N | K x L
+        bag_features = bag_features.view(-1, self.num_encoding * self.pool_out_dim)
+        bag_logit = self.classifier_fn(bag_features)
+        return bag_logit, attentions
 
     def configure_optimizers(self) -> optim.Optimizer:
         return optim.Adam(self.parameters(), lr=self.l_rate, weight_decay=self.weight_decay,